improved data type handling

+++ scrape.py
@@ -53,7 +53,7 @@ def scrape():
             'us_gross':us_gross
         })
 
-        df['year']=df['year'].apply(lambda x:re.findall(r"[0-9]+",x)[0])
+        df['year']=df['year'].apply(lambda x:re.findall(r"[0-9]+",x)[0]).astype(int)
         df['runtime']=df['runtime'].apply(lambda x:re.findall(r"[0-9]+",x)[0] )
         df['votes']=df['votes'].str.replace(',','').astype('int64')
         dfm = dfm.append(df)
